refactor(plan_SERVERSPLIT slice 10): the browser-management bodies leave server.py
The eight `browser-management` tool bodies move VERBATIM out of
embedded/server.py into embedded/tool_sections/browser_management.py. Pure move:
the only edits are the dropped registration decorator (contract rule 2) and the
rewrite of the singleton/knob reads to `rt.<name>`, resolved at CALL time against
the one patchable home (contract rule 3). Nothing else changed — not a docstring,
not a signature, not a comment.

This is the section that owns the BROWSER's own lifecycle, and two of its bodies
are why the mechanism had to be proven on nine easier sections first:

  * spawn_browser is the plan's largest single tool and carries the F-808/F-810
    headed-visibility guard, which runs BEFORE the try and outside it so a spawn
    nobody could ever see refuses without first cloning a profile dir onto disk.
    It reads rt.display_context.display_context() for the refusal message and
    reaches desktop_launch.can_deliver_headed_window() through a FUNCTION-LOCAL
    import, carried as it stood rather than hoisted, which is what keeps
    desktop_launch off this module's import graph.
  * spawn_browser and close_instance are the two ends of the on-disk
    profile/clone lifecycle, so six calls land in rt.clone_storage — resolved at
    call time like every other singleton, which is what keeps
    tests/test_clone_storage.py's "patch it THERE, not on server" pin true of a
    body that no longer lives in server.py.

get_instance_state's `# F-164 non-CDP` marker comment moved byte-unchanged with
it, and tests/test_cdp_timeout.py follows it into the new file through
tests/source_scan.py rather than passing vacuously over an emptier server.py.
That guard is load-bearing exactly here: after this commit server.py contains no
asyncio.wait_for at all.

Census re-run and, as in every slice since the first, ZERO test re-points were
needed: there is no `from ...embedded.server import <name>`, no
`patch("...server.<tool>")` string target and no inspect.getsource anywhere in
tests/ or tools/. tests/test_bug_prone_tools.py's TestSpawnBrowserSubMethodSeam
pins BrowserManager's private pipeline, one layer below the tool body, so the
move cannot reach it. release_evidence.HOST_EXEC_SITES needed nothing this time
— slice 9 took the only tool-body `exec` with it.

Seven imports server.py was the last consumer of are pruned: BrowserOptions, the
two platform_utils predicates, tool_errors._require_landing_ok, and the
tool_runtime aliases display_context, dynamic_hook_system and in_memory_storage
(app_lifespan reads the last of those as rt.*, not through the alias).
clone_storage STAYS, now carrying an explicit keep-reason, because
tests/test_clone_storage.py pins `server.clone_storage is clone_storage` until
slice 12. §5.5's MIGRATION_ALIASES floor RATCHETS DOWN 14 -> 11 accordingly, and
the unit lane legitimately loses one parametrize case per pruned alias.

server.py 1371 -> 986 LOC — under the 1000-LOC default for the first time since
the god file existed; the GRANDFATHER row survives one more slice only because
slice 12 DELETES it rather than merely satisfying it. The cap ratchets DOWN to
the measured post-ruff-format actual, and tests/source_scan.py's floor ratchets
UP 11 -> 12. tool_sections/browser_management.py is 438 LOC.

Verified: wire surface IDENTICAL to tests/goldens/tool_surface.json (94 tools);
SECTION_TOOLS 94 with per-app [94, 94, 94]; `--list-sections` subprocess
Total: 94 tools; unit lane 2326 passed / 1 skipped (base 2329, -3 for the pruned
aliases); real-Chrome test_e2e_transport.py 1 passed, test_browser_integration.py
19 passed, test_clean_shutdown_noise.py 1 skipped / 16 deselected; budgets, ruff
format+check, suppression owners, vulture, ty (96 warnings, 0 errors) and
gen_release_contract --check all clean.

tests/test_stateful_i18n.py reported two failures in this lane, both attributed:
test_storage_and_cookies_survive_one_profile_and_no_other fails identically on
the base with the same stray `_ga` / `_ga_GPM4NSQVSC` cookies in this machine's
persistent browser-session root (recorded in the slices 7-9 log), and
test_service_worker_installs_activates_controls_and_unregisters passes on a
re-run of itself — a load flake on a box carrying 122 live Chrome processes.
